+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed issue with DI using scoped services during debugging in emulator mode
 - Added `PageStyles` and `PageScripts` component to decorate layout
+- Added `MfRouter` component to enable routing (opt-in)
 - Added integration to Piral Inspector protocol for enhanced development
 
 ## 0.2.1 (October 19, 2023)